Move dynamic visit ratio computation into codec layer
Relocates the Two-Signal Model visit ratio calculation from the query
layer (AbstractIVFKnnVectorQuery) into the codec layer (IVFVectorsReader)
where it can leverage per-segment context.

Removes verbose debug logging from IVF search and downgrades segment
layout logging from info to debug level to reduce log noise.

Initializes numCands to k instead of 0 so the dynamic ratio computation
has a meaningful default when no search strategy overrides it.
