+++ server/src/main/java/org/elasticsearch/search/vectors/AbstractIVFKnnVectorQuery.java
@@ -48,14 +48,6 @@ abstract class AbstractIVFKnnVectorQuery extends Query implements QueryProfilerP
 
     static final TopDocs NO_RESULTS = TopDocsCollector.EMPTY_TOPDOCS;
 
-    // Two-Signal Model constants for dynamic visit ratio
-    private static final float V_MIN = 0.003f;
-    private static final float V_MAX = 0.04f;
-    private static final double LOG1P_R_MAX = Math.log1p(10.0);
-    private static final double LOG1P_K_MAX = Math.log1p(10_000.0);
-    private static final double RATIO_WEIGHT = 0.85;
-    private static final double K_WEIGHT = 0.15;
-
     protected final String field;
     protected final float providedVisitRatio;
     protected final int k;
@@ -133,14 +125,9 @@ abstract class AbstractIVFKnnVectorQuery extends Query implements QueryProfilerP
         TaskExecutor taskExecutor = indexSearcher.getTaskExecutor();
         List<LeafReaderContext> leafReaderContexts = reader.leaves();
 
-        final float visitRatio;
-        if (providedVisitRatio == 0.0f) {
-            double r = (double) numCands / Math.max(k, 1);
-            double z = RATIO_WEIGHT * logScale(r - 1.0, LOG1P_R_MAX) + K_WEIGHT * logScale(k, LOG1P_K_MAX);
-            visitRatio = (float) (V_MIN + (V_MAX - V_MIN) * z);
-        } else {
-            visitRatio = providedVisitRatio;
-        }
+        // When providedVisitRatio is 0.0f (dynamic), the codec computes the visit ratio
+        // per-segment using the Two-Signal model with segment-size awareness.
+        final float visitRatio = providedVisitRatio;
 
         List<Callable<TopDocs>> tasks = new ArrayList<>(leafReaderContexts.size());
         for (LeafReaderContext context : leafReaderContexts) {
@@ -210,10 +197,6 @@ abstract class AbstractIVFKnnVectorQuery extends Query implements QueryProfilerP
             knnCollectorManager,
             visitRatio
         );
-    }
-
-    private static double logScale(double value, double log1pMax) {
-        return Math.max(0.0, Math.min(1.0, Math.log1p(value) / log1pMax));
     }
 
     abstract void preconditionQuery(LeafReaderContext context) throws IOException;
